Move <pre> line-comment idempotency tests into shared TCK JavadocTest
The three idempotency cases assert a version-agnostic invariant (round-trip
of // comments inside <pre><code> blocks), so placing them in
rewrite-java-tck exercises every parser version's compatibility suite -
including future ones - rather than only Java 25.
